+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/applicaton/dto/SettingStateDto.kt
@@ -2,4 +2,5 @@ package com.github.naoyukik.intellijplugingithubnotifications.applicaton.dto
 
 data class SettingStateDto(
     val fetchInterval: Int,
+    val repositoryName: String,
 )