+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/domain/SettingStateRepository.kt
@@ -3,5 +3,5 @@ package com.github.naoyukik.intellijplugingithubnotifications.domain
 import com.github.naoyukik.intellijplugingithubnotifications.domain.model.SettingState
 
 interface SettingStateRepository {
-    fun getFetchInterval(): SettingState
+    fun loadSettingState(): SettingState
 }